added restart script

+++ minecraftserver.sh
@@ -28,7 +28,7 @@ export NCURSES_NO_UTF8_ACS=1
 function whitelist() {
 	title="Whitelist"
 	CHOICE=$(dialog --backtitle "$backtitle" --title "$title" \
-	--menu "What would you like to do?" 15 50 2 \
+	--menu "What would you like to do?" 15 50 3 \
 	1 "View players" \
 	2 "Add player" \
 	3 "Remove player" 3>&2 2>&1 1>&3
@@ -36,7 +36,7 @@ function whitelist() {
 
 	case $CHOICE in
 		1) 
-			names=$(cat whitelist.json | grep name)
+			names=$(cat whitelist.json | jq .[].name)
 			listlength=$(echo "$names" | wc -l)
 			if [ ! $listlength = 0 ]; then
 				((listlength+=5))
@@ -51,8 +51,15 @@ function whitelist() {
 			clear
 			playername=$(<"${INPUT}")
 			id=$(GET https://api.mojang.com/users/profiles/minecraft/$playername)
-			echo $id
+			names=$(cat whitelist.json | jq .[].name)
+			listlength=$(echo "$names" | wc -l)
+			jq ".[$listlength] |= .+$id" whitelist.json | tee whitelist.json
+			sed -i 's/"id"/"uuid"/g' whitelist.json
+			read -p "Holding..."
+			dialog --backtitle "$backtitle" --title "$title" \
+				--msgbox "Added $id to whitelist" 8 40
 			exit
+
 			;;
 		3) echo whoa there;;
 	esac	
@@ -142,27 +149,59 @@ function info() {
 }
 
 function start() {
-	# Display Minecraft and Java version then pause
-	echo -e "$servertype version $mcver with maximum $maxram and minimum $minram of RAM"
-	java -version
-	echo
-	echo -e "\e[93m[Ctrl+A] \e[96mthen \e[93m[D] \e[96mto detach"
-	echo -e "type \e[93m'minecraftserver' \e[96mto reattatch\e[0m"
-	echo
-	read -p "Press [Enter] to start the server with the current settings..."
+	while true
+	do
+		# Display Minecraft and Java version then pause
+		echo -e "$servertype version $mcver with maximum $maxram and minimum $minram of RAM"
+		java -version
+		echo
+		echo -e "\e[93m[Ctrl+A] \e[96mthen \e[93m[D] \e[96mto detach"
+		echo -e "type \e[93m'minecraftserver' \e[96mto reattatch\e[0m"
+		echo
+	#	read -p "Press [Enter] to start the server with the current settings..."
+	
+		# Start the Server with minram and maxram values
+		java -Xmx$maxram -Xms$minram -jar $servertype-$mcver.jar
+		#read -p "server up!" yeetus
+		echo -e "\e[91mServer stopped!\e[0m"
+		source "$serverdir/restart"
+		if [[ $RESTART = "1" ]]; then
 
-	# Start the Server with minram and maxram values
-	java -Xmx$maxram -Xms$minram -jar $servertype-$mcver.jar
-	echo -e "\e[91mServer stopped!\e[0m"
-	read -p "Press [Enter] to close this window..."
-	echo "bye bye!"
-	sleep 1
-	tmux kill-session -t minecraft
+			rm $serverdir/restart
+			RESTART=0
+			echo "Press 'Q' to cancel restart"
+			echo "Restarting in..."
+			for i in 5 4 3 2 1
+			do
+				echo "$i..."
+				
+				read -t 1 -N 1 input
+				if [[ $input = "q" ]] || [[ $input = "Q" ]]; then
+					# The following line is for the prompt to appear on a new line.
+			        	echo
+				#	read -p "Press [Enter] to close this window..."
+					echo "bye bye!"
+					sleep 1
+					tmux kill-session -t minecraft
+					exit
+	
+				fi
+			done
+			echo "Rebooting now!"
+		else	
+			echo "Bye bye!"
+			sleep 1
+			tmux kill-session -t minecraft
+		fi
+	done
+	
+
 }
 
 function initialize() {
 	tmux new-session -s minecraft -d 'minecraftserver forcestart'
-	tmux split-window -h 'minecraftserver term'
+	#tmux split-window -h 'minecraftserver term'
+	tmux split-window -h 'gtop'
 	tmux split-window -v 'minecraftserver info'
 	tmux select-pane -L 
 	tmux attach -t minecraft
@@ -190,6 +229,7 @@ check
 [ $1 = info ] && info
 [ $1 = term ] && term
 [ $1 = whitelist ] && whitelist
+[ $1 = init ] && initialize
 
 # Test to see if tmux is installed
 if ! [ -x "$(command -v tmux)" ]; then
